requirements.txt
pandas==0.24.2
pyspark==3.0.0
numpy==1.16.4
click==7.1.2
pyyaml==5.3.1
jupyter==1.0.0
scikit-learn==0.23.1
matplotlib==2.2.5
relieff==0.1.2
Cython==0.29.21
pymrmr==0.1.8
skrebate==0.61
xgboost==0.82
tensorflow==2.0.0b0
keras==2.3.1
imblearn==0.0
pyyaml==5.3.1